moved CreateBlobServiceClient to StorageAccountHelpers
